+++ core/src/main/java/Gameplay/Movimientos/IMovimiento.java
@@ -1,7 +1,3 @@
 package Gameplay.Movimientos;
 
-import entidades.personajes.Personaje;
-
-public interface IMovimiento {
-    void ejecutar(Personaje personaje);
-}
+public interface IMovimiento { }