+++ core/src/main/java/entidades/proyectiles/Roca.java
@@ -10,6 +10,8 @@ import managers.GestorAssets;
 
 public class Roca extends Proyectil {
 
+    private int radioDestruccion = 50;
+
     public Roca(float x, float y, float angulo, float velocidad, int dano, GestorColisiones gestorColisiones, Personaje ejecutor) {
         super(x, y, angulo, velocidad, dano, gestorColisiones, ejecutor);
         this.textura = GestorAssets.get("roca.png", Texture.class);
@@ -28,10 +30,15 @@ public class Roca extends Proyectil {
     }
 
     @Override
-    public void impactoMapa(Mapa mapa) {
-        int centroX = (int) (x + hitbox.width / 2f);
-        int centroY = mapa.getHeight() - 1 - (int) (y + hitbox.height / 2f);
-        mapa.destruir(x + hitbox.width/2f, y + hitbox.height/2f, 10);
-        System.out.println("Impacto con el mapa en (" + centroX + ", " + centroY + ")");
+    public void impactoMapa(Mapa mapa, int radio) {
+        float centroXMundo = x + hitbox.getWidth() / 2f;
+        float centroYMundo = y + hitbox.getHeight() / 2f;
+
+        mapa.destruir(centroXMundo, centroYMundo, this.radioDestruccion);
+
+        this.activo = false;
     }
+
+    public int getRadioDestruccion() { return radioDestruccion; }
+
 }